changed datatype to int for deptID

+++ src/com/java/Department.java
@@ -5,7 +5,7 @@ public class Department {
 public static void main(String[] args) {
 
 	String deptName;
-	double deptID;
+	int deptID;
 	String Section;
 	String mgr;
 	//change3
